% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean_cer.R
\name{balance_kwh}
\alias{balance_kwh}
\title{creates a balanced panel of CER data.}
\usage{
balance_kwh(DT_KWH, drop_dst = TRUE)
}
\arguments{
\item{DT_KWH}{(data.table) table with CER consumption data}

\item{drop_dst}{(logical) drop Daylight savings observations (data will NOT be perfectly balance if FALSE)}
}
\description{
creates a balanced panel of CER data.
}

