% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean_cer.R
\name{find_seq_zeros_nas}
\alias{find_seq_zeros_nas}
\title{drop ids with sequential zero readings.}
\usage{
find_seq_zeros_nas(DT_KWH, tol = 6, drop = TRUE)
}
\arguments{
\item{tol}{tolerable sequence of zeros and missing vaues}
}
\description{
drop ids with sequential zero readings.
}

