% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean_cer.R
\name{imput_residential_kwh_days}
\alias{imput_residential_kwh_days}
\title{imput residential kwh data. imputation can take a LONG time.}
\usage{
imput_residential_kwh_days(DT_KWH)
}
\description{
imput residential kwh data. imputation can take a LONG time.
}

