% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean_cer.R
\name{imput_residential_kwh_time}
\alias{imput_residential_kwh_time}
\title{imput residential kwh zeros and missing data along time. imputation can take a LONG time.}
\usage{
imput_residential_kwh_time(DT_KWH)
}
\description{
imput residential kwh zeros and missing data along time. imputation can take a LONG time.
}

