% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean_cer.R
\name{impute_residential_kwh_time}
\alias{impute_residential_kwh_time}
\title{IN TESTING
impute residential kwh zeros and missing data along time. imputation can take a LONG time.}
\usage{
impute_residential_kwh_time(DT_KWH)
}
\description{
IN TESTING
impute residential kwh zeros and missing data along time. imputation can take a LONG time.
}

