% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/panel_data_functions.R
\name{tsls}
\alias{tsls}
\title{Run Two Stage Least Squares}
\usage{
tsls(d, z, x, y)
}
\arguments{
\item{d}{endogenous variable (vector)}

\item{z}{instrument (vector or matrix)}

\item{x}{exogenous variable (vector or matrix)}

\item{y}{dependent variable (vector)}
}
\value{
Returns a linear regression model `lm` where the endogenous variable has been replaced by the predicted values of the first stage. That is, the results of a two stage lease squares regression.
}
\description{
Run Two Stage Least Squares
}

