% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean_cer.R
\name{balance_kwh}
\alias{balance_kwh}
\title{Useful functions for cleaning problems with the CER data. Used AFTER one has imported
the raw data using get_cer(). It will first fill in any}
\usage{
balance_kwh(DT_KWH, drop_dst = FALSE)
}
\arguments{
\item{DT_KWH}{(data.table) table with CER consumption data}

\item{drop_dst}{(logical) drop Daylight savings observations (data will NOT be perfectly balance if FALSE)}
}
\description{
Useful functions for cleaning problems with the CER data. Used AFTER one has imported
the raw data using get_cer(). It will first fill in any
}

