% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean_cer.R
\name{find_seq_zeros_nas}
\alias{find_seq_zeros_nas}
\title{drop ids with long sequential zero readings or missing values.}
\usage{
find_seq_zeros_nas(DT_KWH, tol = 10, drop = TRUE)
}
\arguments{
\item{tol}{tolerable sequence of zeros and missing values}

\item{drop}{drops ids with long sequence of zeroes}
}
\description{
drop ids with long sequential zero readings or missing values.
}

